+++ src/main/Item.java
@@ -149,7 +149,7 @@ abstract public class Item implements Purchasable {
      * @return the string representing what the player sold
      */
     public String sell() throws NotFoundException, InvalidValueException {
-    	player.addBalance(cost * refundAmount);
+    	player.addBalance((int) (cost * refundAmount));
     	player.getItems().remove(this);
     	return "You sold: " + name;
     }
